+++ coding/codemirror/logtalk.js
@@ -534,7 +534,7 @@ CodeMirror.defineMode("logtalk", function(cmCfg, modeCfg) {
              { token: 'support.function.control.logtalk',
                regex: '\\b(?:true|fa(?:il|lse)|repeat|(?:instantiation|system)_error)\\b(?![-!(^~])' },
              { token: 'support.function.control.logtalk',
-               regex: '\\b((?:type|domain|existence|permission|representation|evaluation|resource|syntax)_error)(?=[(])' },
+               regex: '\\b((?:uninstantiation|type|domain|existence|permission|representation|evaluation|resource|syntax)_error)(?=[(])' },
              { token: 'support.function.control.logtalk',
                regex: '\\b(?:ca(?:ll|tch)|ignore|throw|once)(?=[(])' },
              { token: 'support.function.chars-and-bytes-io.logtalk',