+++ coding/rouge/lib/rouge/lexers/logtalk.rb
@@ -62,7 +62,7 @@ module Rouge
         # Control constructs
         rule /(call|catch|throw)(?=[(])/, Keyword
         rule /(fail|false|true|instantiation_error|system_error)\b/, Keyword
-        rule /(type_error|domain_error|existence_error|permission_error|representation_error|evaluation_error|resource_error|syntax_error)(?=[(])/, Keyword
+        rule /(uninstantiation_error|type_error|domain_error|existence_error|permission_error|representation_error|evaluation_error|resource_error|syntax_error)(?=[(])/, Keyword
         # All solutions
         rule /(bagof|setof|findall|forall)(?=[(])/, Keyword
         # Multi-threading meta-predicates