+++ coding/syntaxhighlighter/shBrushLogtalk.js
@@ -5,7 +5,7 @@
 
 	function Brush()
 	{
-		// Contributed by Paulo Moura (https://logtalk.org/); last revised on February 12, 2019
+		// Contributed by Paulo Moura (https://logtalk.org/); last revised on July 31, 2020
 
 		this.regexList = [
 		  // variables
@@ -96,7 +96,7 @@
 			{ regex: new RegExp("\\b(ca(ll|tch)|throw)(?=[(])", 'g'), css: 'keyword'},
 			{ regex: new RegExp("\\b(true|fa(il|lse)|repeat|(instantiation|system)_error)\\b", 'g'), css: 'keyword'},
 			{ regex: new RegExp("\\b(ignore|once)(?=[(])", 'g'), css: 'keyword'},
-			{ regex: new RegExp("\\b(type|domain|existence|permission|representation|evaluation|resource|syntax)_error(?=[(])", 'g'), css: 'keyword'},
+			{ regex: new RegExp("\\b(uninstantiation|type|domain|existence|permission|representation|evaluation|resource|syntax)_error(?=[(])", 'g'), css: 'keyword'},
 			// directives
 			{ regex: new RegExp("^\\s*:\-\\s(c(a(lls|tegory)|oinductive)|p(ublic|r(ot(ocol|ected)|ivate))|e(l(if|se)|n(coding|sure_loaded)|xport)|i(f|n(clude|itialization|fo))|alias|d(ynamic|iscontiguous)|m(eta_(non_terminal|predicate)|od(e|ule)|ultifile)|reexport|s(et_(logtalk|prolog)_flag|ynchronized)|o(bject|p)|use(s|_module))(?=[(])", 'gm'), css: 'preprocessor' },
 			{ regex: new RegExp("^\\s*:\-\\s(e(lse|nd(if|_(category|object|protocol)))|built_in|dynamic|synchronized|threaded)\.", 'gm'), css: 'preprocessor' },